app/utils/config.py
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Google Gemini API key
GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")
if not GOOGLE_API_KEY:
    raise ValueError("GOOGLE_API_KEY environment variable is not set")

# News API key
NEWSAPI_KEY = os.getenv("NEWSAPI_KEY")
if not NEWSAPI_KEY:
    raise ValueError("NEWSAPI_KEY environment variable is not set") 