requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
yfinance==0.2.36
python-dotenv==1.0.1
pydantic==2.6.1
google-generativeai==0.3.2
httpx==0.26.0
python-dateutil==2.8.2
newsapi-python==0.2.7
pandas==2.2.0 